workshop/requirements.txt
# LangChain/LangGraph Workshop Requirements
# Compatible with LangChain v1.0+

# Core LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# LLM Providers
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0  # Optional: if using Claude

# LangGraph
langgraph>=0.0.20

# Vector Stores & Embeddings
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment if have CUDA

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional but recommended
langsmith>=0.0.70  # For monitoring and tracing
tiktoken>=0.5.0    # Token counting
